pyproject.toml
[tool.poetry]
name = "pin-drop-sampling2"
version = "0.1.0"
description = ""
authors = ["dougj892 <dougj892@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.13,>=3.11"
geopandas = "^1.0.1"
pyarrow = "^18.0.0"
folium = "^0.18.0"
s2sphere = "^0.2.5"
tqdm = "^4.66.6"
osmnx = {version = "2.0.0b3", allow-prereleases = true}
python-dotenv = "^1.0.1"
geopy = "^2.4.1"
openpyxl = "^3.1.5"
matplotlib = "^3.9.2"
scipy = "^1.14.1"
statsmodels = "^0.14.4"
fiona = "^1.10.1"
simplekml = "^1.3.6"
ipykernel = "^6.29.5"
rpy2 = "^3.5.17"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
